frontend_api/requirements.txt
Django
django_debug_toolbar
djangorestframework
djangorestframework-simplejwt
dj-database-url
python-decouple
django-cors-headers
whitenoise
django-filter
pytest
pytest-django
factory-boy
pytest-mock
Faker
django-cloudinary-storage
cloudinary
Pillow
drf-spectacular
python-dotenv
gunicorn
redis
